Safari-609.1.11.1/LayoutTests/fast/dom/Window/window-open-self-as-opener-expected.txt

Tests navigating a window whose opener is itself

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS w.opener is self
PASS w.opener is w
PASS w.opener is self
PASS successfullyParsed is true

TEST COMPLETE
